ADAS-34-feat/tc275_acc_lkas_개발

+++ code/controller/ap/core/car_distance.c
@@ -54,6 +54,6 @@ void set_car_distance(int count){
     adas_controller.adas_sensor.ultrasonic[count] = ultrasonic_read_distance(count);
 }
 
-int get_car_distance(int count){
+float get_car_distance(int count){
     return adas_controller.adas_sensor.ultrasonic[count];
 }
